Переносит существующие тесты в редьюсер app
