+++ src/utils/common/common.test.js
@@ -1,34 +1,8 @@
 import {
-  filterMoviesByGenre,
-  getGenres,
   getRatingTextDescription,
   transformToFirstCapitalSymbol
 } from "./common";
-import {ALL_GENRES_NAME, TextRating} from "../../const";
-
-const movies = [
-  {
-    genre: [`genre1`]
-  },
-  {
-    genre: [`genre1`, `genre2`]
-  },
-  {
-    genre: [`genre1`, `genre2`, `genre3`]
-  },
-  {
-    genre: [`genre3`]
-  },
-  {
-    genre: [`genre4`]
-  },
-  {
-    genre: [`genre5`]
-  },
-  {
-    genre: [`genre5`]
-  }
-];
+import {TextRating} from "../../const";
 
 describe(`Check the rating is correct`, () => {
   it(`should return correct rating with boundary value`, () => {
@@ -67,63 +41,5 @@ describe(`Check the string transform is correct`, () => {
 
   it(`should return correct string with uppercase string`, () => {
     expect(transformToFirstCapitalSymbol(`TEST`)).toBe(`Test`);
-  });
-});
-
-describe(`Check the filtered movies is correct`, () => {
-  it(`should return movies with unique genre`, () => {
-    expect(filterMoviesByGenre(movies, `genre4`))
-      .toEqual([
-        {
-          genre: [`genre4`]
-        }
-      ]);
-  });
-
-  it(`should return movies with not unique genre`, () => {
-    expect(filterMoviesByGenre(movies, `genre1`))
-      .toEqual([
-        {
-          genre: [`genre1`]
-        },
-        {
-          genre: [`genre1`, `genre2`]
-        },
-        {
-          genre: [`genre1`, `genre2`, `genre3`]
-        }
-      ]);
-  });
-
-  it(`should return movies with same genre`, () => {
-    expect(filterMoviesByGenre(movies, `genre5`))
-      .toEqual([
-        {
-          genre: [`genre5`]
-        },
-        {
-          genre: [`genre5`]
-        }
-      ]);
-  });
-
-  it(`should return no movies with nonexistent genre`, () => {
-    expect(filterMoviesByGenre(movies, `nonexistent genre`))
-      .toEqual([]);
-  });
-
-  it(`should return all movies with all genres`, () => {
-    expect(filterMoviesByGenre(movies, ALL_GENRES_NAME))
-      .toEqual(movies);
-  });
-});
-
-describe(`Check creating a list of genres`, () => {
-  it(`should return correct genres list`, () => {
-    expect(getGenres(movies)).toEqual([ALL_GENRES_NAME, `genre1`, `genre2`, `genre3`, `genre4`, `genre5`]);
-  });
-
-  it(`should return correct genres list with empty movies`, () => {
-    expect(getGenres([])).toEqual([ALL_GENRES_NAME]);
   });
 });
